+++ client/src/components/hunting/HuntingDropDown.tsx
@@ -1,25 +1,9 @@
 import React from "react";
 import styled from "styled-components";
-import useDetectClose from "../util/useDetectClose";
+import useDetectClose from "../../util/useDetectClose";
 import { useDispatch, useSelector } from "react-redux";
-import { RootState } from "../util/store";
-import { selectCategory, selectLocation, selectTag } from "../util/types";
-
-const tags = [
-  "재택근무🏠",
-  "야간🌙",
-  "초보자가능🐣",
-  "최저시급💰",
-  "당일지급💵",
-  "능력활용🧐",
-  "역세권🚇",
-  "식사제공🍴",
-  "경력1년이상💡",
-];
-
-const TagButton = ({ tag }: { tag: string; onClick: () => void }) => (
-  <button>{tag}</button>
-);
+import { RootState } from "../../util/store";
+import { selectCategory, selectLocation } from "../../util/types";
 
 const category = [
   "카테고리",
@@ -66,7 +50,7 @@ const location = [
   "강동구",
 ];
 
-const DropdownMenu = () => {
+const DropdownT = () => {
   const dispatch = useDispatch();
   const selectedCategory = useSelector(
     (state: RootState) => state.selectedCategory
@@ -74,7 +58,6 @@ const DropdownMenu = () => {
   const selectedLocation = useSelector(
     (state: RootState) => state.selectedLocation
   );
-  const selectedTag = useSelector((state: RootState) => state.selectedTag);
 
   const handleCategoryClick = (category: string) => {
     dispatch(selectCategory(category));
@@ -86,14 +69,10 @@ const DropdownMenu = () => {
     locationHandler();
   };
 
-  const handleTagClick = (tag: string) => {
-    dispatch(selectTag(tag));
-  };
-
   const [categoryIsOpen, categoryRef, categoryHandler] = useDetectClose(false);
   const [locationIsOpen, locationRef, locationHandler] = useDetectClose(false);
 
-  const newHireClickHandler = () => {
+  const newHuntingClickHandler = () => {
     console.log("새 글 작성");
   };
 
@@ -102,9 +81,7 @@ const DropdownMenu = () => {
       <UpperWrapper>
         <DropdownContainer>
           <DropdownWrapper
-            onClick={() => handleCategoryClick("category")}
-            ref={categoryRef}
-          >
+            onClick={categoryHandler} ref={categoryRef}>
             <span>{selectedCategory}</span>
           </DropdownWrapper>
           <DropdownTitle isDropped={categoryIsOpen}>
@@ -112,7 +89,6 @@ const DropdownMenu = () => {
               {category.map((category: string) => (
                 <DropdownItem key={category}>
                   <LinkWrapper
-                    href="#1-1"
                     onClick={() => handleCategoryClick(category)}
                   >
                     {category}
@@ -132,7 +108,6 @@ const DropdownMenu = () => {
               {location.map((location: string) => (
                 <DropdownItem key={location}>
                   <LinkWrapper
-                    href="#2-1"
                     onClick={() => handleLocationClick(location)}
                   >
                     {location}
@@ -143,30 +118,10 @@ const DropdownMenu = () => {
           </DropdownTitle>
         </DropdownContainer>
 
-        <AddHire onClick={newHireClickHandler}>
+        <AddHire onClick={newHuntingClickHandler}>
           <button>게시글 작성</button>
         </AddHire>
       </UpperWrapper>
-      <LowerWrapper>
-        <CategoryButton>
-          <div>
-            {tags.map((tag) => (
-              <TagButton
-                key={1}
-                tag={tag}
-                onClick={() => handleTagClick(selectedTag)}
-              />
-            ))}
-          </div>
-        </CategoryButton>
-        <FilterButton>
-          <div>
-            <button>조회순</button>
-            <button>보수높은순</button>
-            <button>최신순</button>
-          </div>
-        </FilterButton>
-      </LowerWrapper>
     </>
   );
 };
@@ -273,27 +228,4 @@ const AddHire = styled.div`
   font-size: 16px;
 `;
 
-//태그 파트
-const LowerWrapper = styled.div`
-  width: 100%;
-  height: 100px;
-  padding: 20px 20px 15px 20px;
-  border-bottom: 1px solid #dadbdc;
-`;
-
-const CategoryButton = styled.button`
-  margin: 16px 50px 8px;
-  border: none;
-  background-color: white;
-`;
-
-const FilterButton = styled.div`
-  height: 38px;
-  margin: 0 20px 0 8px;
-  float: right;
-  border: none;
-  cursor: pointer;
-  font-size: 16px;
-`;
-
-export default DropdownMenu;
+export default DropdownT;
